usb: musb: allow for post fs_initcall binding
At initialization, the musb driver does a platform_driver_probe to attach to
the device in question.  As a result, it fails to bind if the platform_device
is not already registered by then.

With this patch the driver binds properly even if the device is added at a
later point.