musb: cppi41: remote autoreq and mode initialization
AUTOREQ and MODE registers are configured within the data path on a
per-transfer basis.  These do not need to be done at controller
initialization.  Doing so causes an undesirable mix of CPPI and MUSB register
accesses, and is problematic when multiple musb controllers are attached to
a shared cppi dma instance.